Configure webpack for css prefixes and minizing
Adds dev dependencies autoprefixer, css-loader, css-minimizer-webpack-plugin,
mini-css-extract-plugin, postcss, and postcss-loader. Also moves all
dependencies into dev dependencies, as webpack's purpose in this project is to
build files.

In webpack.config.js, the entry point is modified to include main.css. Also,
the output path is specified to just be the dist/ folder and output files now
include the file path. The MiniCssExtractPlugin loader, css-loader, and
postcss-loader are all used for css files with the CssMinimizerPlugin applied
as an optimizing minifier.

To use postcss, postcss.config.js is created and uses the autoprefixer plugin
to apply browser prefixes to relevant css attributes.